fix: use  for  params on Windows

+++ src-tauri/libvips-win/src/lib.rs
@@ -43,7 +43,7 @@ impl VipsApp {
         }
     }
     
-    pub fn get_disc_threshold(&self) -> u64 {
+    pub fn get_disc_threshold(&self) -> usize {
         unsafe { bindings::vips_get_disc_threshold() }
     }
     
@@ -126,7 +126,7 @@ impl VipsApp {
         }
     }
 
-    pub fn cache_set_max_mem(&self, max: u64) {
+    pub fn cache_set_max_mem(&self, max: usize) {
         unsafe {
             bindings::vips_cache_set_max_mem(max);
         }
@@ -138,7 +138,7 @@ impl VipsApp {
         }
     }
 
-    pub fn cache_get_max_mem(&self) -> u64 {
+    pub fn cache_get_max_mem(&self) -> usize {
         unsafe {
             bindings::vips_cache_get_max_mem()
         }
@@ -188,13 +188,13 @@ impl VipsApp {
         }
     }
 
-    pub fn tracked_get_mem(&self) -> u64 {
+    pub fn tracked_get_mem(&self) -> usize {
         unsafe {
             bindings::vips_tracked_get_mem()
         }
     }
 
-    pub fn tracked_get_mem_highwater(&self) -> u64 {
+    pub fn tracked_get_mem_highwater(&self) -> usize {
         unsafe {
             bindings::vips_tracked_get_mem_highwater()
         }
