+++ src-tauri/libvips-win/src/ops.rs
@@ -9389,7 +9389,7 @@ pub fn ppmload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_ppmload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_ppmload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -9472,7 +9472,7 @@ pub fn ppmload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_ppmload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             flags_in_name.as_ptr(),
             flags_in,
@@ -9720,7 +9720,7 @@ pub fn radload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_radload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_radload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -9803,7 +9803,7 @@ pub fn radload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_radload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             flags_in_name.as_ptr(),
             flags_in,
@@ -10102,7 +10102,7 @@ pub fn svgload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_svgload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_svgload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -10231,7 +10231,7 @@ pub fn svgload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_svgload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             dpi_in_name.as_ptr(),
             dpi_in,
@@ -10398,7 +10398,7 @@ pub fn gifload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_gifload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_gifload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -10495,7 +10495,7 @@ pub fn gifload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_gifload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             n_in_name.as_ptr(),
             n_in,
@@ -10774,7 +10774,7 @@ pub fn pngload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_pngload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_pngload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -10868,7 +10868,7 @@ pub fn pngload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_pngload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             unlimited_in_name.as_ptr(),
             unlimited_in,
@@ -11158,7 +11158,7 @@ pub fn jpegload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_jpegload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_jpegload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -11270,7 +11270,7 @@ pub fn jpegload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_jpegload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             shrink_in_name.as_ptr(),
             shrink_in,
@@ -11439,7 +11439,7 @@ pub fn webpload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_webpload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_webpload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -11543,7 +11543,7 @@ pub fn webpload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_webpload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             page_in_name.as_ptr(),
             page_in,
@@ -11869,7 +11869,7 @@ pub fn tiffload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_tiffload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_tiffload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -11995,7 +11995,7 @@ pub fn tiffload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_tiffload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             page_in_name.as_ptr(),
             page_in,
@@ -12342,7 +12342,7 @@ pub fn heifload_buffer(buffer: &[u8]) -> Result<VipsImage> {
         let mut out_out: *mut bindings::VipsImage = null_mut();
 
         let vips_op_response =
-            bindings::vips_heifload_buffer(buffer_in, buffer.len() as u64, &mut out_out, NULL);
+            bindings::vips_heifload_buffer(buffer_in, buffer.len(), &mut out_out, NULL);
         utils::result(
             vips_op_response,
             || VipsImage { ctx: out_out },
@@ -12461,7 +12461,7 @@ pub fn heifload_buffer_with_opts(
 
         let vips_op_response = bindings::vips_heifload_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             page_in_name.as_ptr(),
             page_in,
@@ -13271,7 +13271,7 @@ pub fn rawsave_with_opts(
 pub fn rawsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -13327,7 +13327,7 @@ pub fn rawsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let keep_in: i32 = rawsave_buffer_options.keep as i32;
@@ -14035,7 +14035,7 @@ pub fn radsave_with_opts(
 pub fn radsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -14091,7 +14091,7 @@ pub fn radsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let keep_in: i32 = radsave_buffer_options.keep as i32;
@@ -14417,7 +14417,7 @@ pub fn gifsave_with_opts(
 pub fn gifsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -14505,7 +14505,7 @@ pub fn gifsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let dither_in: f64 = gifsave_buffer_options.dither;
@@ -14960,7 +14960,7 @@ pub fn pngsave_with_opts(
 pub fn pngsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -15053,7 +15053,7 @@ pub fn pngsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let compression_in: i32 = pngsave_buffer_options.compression;
@@ -15523,7 +15523,7 @@ pub fn jpegsave_with_opts(
 pub fn jpegsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -15617,7 +15617,7 @@ pub fn jpegsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let q_in: i32 = jpegsave_buffer_options.q;
@@ -16375,7 +16375,7 @@ pub fn webpsave_with_opts(
 pub fn webpsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -16496,7 +16496,7 @@ pub fn webpsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let q_in: i32 = webpsave_buffer_options.q;
@@ -17444,7 +17444,7 @@ pub fn tiffsave_with_opts(
 pub fn tiffsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -17598,7 +17598,7 @@ pub fn tiffsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let compression_in: i32 = tiffsave_buffer_options.compression as i32;
@@ -18281,7 +18281,7 @@ pub fn heifsave_with_opts(
 pub fn heifsave_buffer(inp: &VipsImage) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let vips_op_response =
@@ -18377,7 +18377,7 @@ pub fn heifsave_buffer_with_opts(
 ) -> Result<Vec<u8>> {
     unsafe {
         let inp_in: *mut bindings::VipsImage = inp.ctx;
-        let mut buffer_buf_size: u64 = 0;
+        let mut buffer_buf_size: usize = 0;
         let mut buffer_out: *mut c_void = null_mut();
 
         let q_in: i32 = heifsave_buffer_options.q;
@@ -18849,7 +18849,7 @@ pub fn thumbnail_buffer(buffer: &[u8], width: i32) -> Result<VipsImage> {
 
         let vips_op_response = bindings::vips_thumbnail_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             width_in,
             NULL,
@@ -18995,7 +18995,7 @@ pub fn thumbnail_buffer_with_opts(
 
         let vips_op_response = bindings::vips_thumbnail_buffer(
             buffer_in,
-            buffer.len() as u64,
+            buffer.len(),
             &mut out_out,
             width_in,
             option_string_in_name.as_ptr(),
